feat(app): v1.2.0 Refactor app codebase
