+++ src/dataStore.js
@@ -1,0 +1,8 @@
+import {CELSIUS_UNITS} from "./utils";
+
+const dataStore = {
+  currentCity: '',
+  currentUnits: CELSIUS_UNITS,
+};
+
+export default dataStore;
