+++ src/framework/render.js
@@ -1,0 +1,9 @@
+import App from "../components/App";
+
+function renderApp() {
+  document.getElementById('app-root').innerHTML = `
+        ${App()}
+    `;
+}
+
+export default renderApp;
