+++ src/index.js
@@ -1,0 +1,11 @@
+import renderApp from "./framework/render";
+import dataStore from "./dataStore";
+
+if (module.hot) {
+  module.hot.accept();
+}
+
+window.renderApp = renderApp;
+window.dataStore = window.dataStore || dataStore;
+
+renderApp();
